Refactor extractor classes
