+++ catalogue_graph/src/ingestor/models/display/image.py
@@ -1,7 +1,7 @@
 from pydantic import computed_field
 
-from ingestor.extractors.images_extractor import ExtractedImage
-from ingestor.transformers.work_display_transformer import DisplayWorkDataTransformer
+from ingestor.extractors.images.images_extractor import ExtractedImage
+from ingestor.transformers.work_display_transformer import DisplayWorkTransformer
 from models.pipeline.serialisable import ElasticsearchModel
 
 from .concept import DisplayContributor, DisplayGenre, DisplaySubject
@@ -16,13 +16,13 @@ class DisplayImageSource(ElasticsearchModel):
     subjects: list[DisplaySubject]
     genres: list[DisplayGenre]
     languages: list[DisplayIdLabel]
-    type: str = "oork"
+    type: str = "Work"
 
     @classmethod
     def from_extracted_image(cls, extracted: ExtractedImage) -> "DisplayImageSource":
         image = extracted.image
         work_data = image.source.data
-        transformer = DisplayWorkDataTransformer(work_data, extracted.concepts)
+        transformer = DisplayWorkTransformer(extracted.work)
 
         return DisplayImageSource(
             id=image.source.id.canonical_id,
@@ -42,7 +42,8 @@ class DisplayImage(ElasticsearchModel):
     source: DisplayImageSource
     type: str = "Image"
 
-    @computed_field
+    # These stacked decorators are supported but mypy thinks they are not
+    @computed_field  # type: ignore
     @property
     def thumbnail(self) -> DisplayDigitalLocation:
         for location in self.locations:
@@ -59,7 +60,8 @@ class DisplayImage(ElasticsearchModel):
         return DisplayImage(
             id=image.source.id.canonical_id,
             locations=[
-                DisplayDigitalLocation.from_location(loc) for loc in image.locations
+                DisplayDigitalLocation.from_digital_location(loc)
+                for loc in image.locations
             ],
             aspect_ratio=inferred_data.aspect_ratio or 1,
             average_color=inferred_data.average_color_hex or "#ffffff",
