+++ catalogue_graph/src/ingestor/models/filter/work.py
@@ -1,6 +1,6 @@
 from pydantic import Field
 
-from ingestor.extractors.works_extractor import VisibleExtractedWork
+from ingestor.extractors.works.base_works_extractor import VisibleExtractedWork
 from ingestor.transformers.work_query_transformer import QueryWorkTransformer
 from models.pipeline.serialisable import ElasticsearchModel
