+++ catalogue_graph/src/ingestor/models/query/image.py
@@ -1,4 +1,4 @@
-from ingestor.extractors.images_extractor import ExtractedImage
+from ingestor.extractors.images.images_extractor import ExtractedImage
 from models.pipeline.serialisable import ElasticsearchModel
 
 from .work import QueryWork
@@ -10,13 +10,7 @@ class QueryImage(ElasticsearchModel):
 
     @classmethod
     def from_extracted_image(cls, extracted: ExtractedImage) -> "QueryImage":
-        source_work = extracted.image.source
         return QueryImage(
             id=extracted.image.state.canonical_id,
-            source=QueryWork.from_work_data(
-                data=source_work.data,
-                concepts=extracted.concepts,
-                work_id=source_work.id.canonical_id,
-                source_identifier=source_work.id.source_identifier,
-            ),
+            source=QueryWork.from_extracted_work(extracted.work),
         )
