+++ catalogue_graph/src/ingestor/steps/ingestor_loader.py
@@ -59,6 +59,9 @@ def handler(
     )
 
     es_client = get_client("read_only", event.pipeline_date, es_mode)
+    # es_client = get_client(
+    #     f"{event.ingestor_type}_ingestor", event.pipeline_date, es_mode
+    # )
     neptune_client = NeptuneClient(event.environment)
     transformer = create_transformer(event, es_client, neptune_client)
     objects_to_index = transformer.load_documents(event, load_destination)
