+++ catalogue_graph/tests/ingestor/test_works_extractor.py
@@ -3,10 +3,10 @@ from typing import Literal
 
 import pytest
 
-from ingestor.extractors.works_extractor import (
-    GraphWorksExtractor,
+from ingestor.extractors.works.base_works_extractor import (
     VisibleExtractedWork,
 )
+from ingestor.extractors.works.works_index_extractor import WorksIndexExtractor
 from ingestor.models.merged.work import VisibleMergedWork
 from ingestor.models.neptune.query_result import (
     ExtractedConcept,
@@ -33,8 +33,8 @@ EXTRACTED_CONCEPT_FIXTURE = load_json_fixture("neptune/full_extracted_concept.js
 EXPECTED_EXTRACTED_CONCEPT = ExtractedConcept.model_validate(EXTRACTED_CONCEPT_FIXTURE)
 
 
-def get_extractor() -> GraphWorksExtractor:
-    return GraphWorksExtractor(
+def get_extractor() -> WorksIndexExtractor:
+    return WorksIndexExtractor(
         MOCK_EVENT,
         get_mock_es_client("graph_extractor", MOCK_EVENT.pipeline_date),
         get_mock_neptune_client(),
@@ -69,7 +69,7 @@ def mock_graph_relationships(
     if "concepts" in include:
         concept_ids = [MOCK_CONCEPT_ID]
         monkeypatch.setattr(
-            "ingestor.extractors.works_extractor.extract_identified_concept_ids",
+            "ingestor.extractors.works.base_works_extractor.extract_identified_concept_ids",
             lambda _: concept_ids,
         )
         add_mock_responses_for_ids([MOCK_CONCEPT_ID])
